Add files via upload
# Financial Performance Analysis

### **Situation**

At a leading consulting firm, the finance team is facing a complex challenge: the company lacks a centralised view of the financial perfomance and financial models are created at a country level without being aggregated. This lack of clarity obscured the true financial health and performance across different services and regions, making it difficult to:

- Understand the revenue distribution and profitability by country and service line.
- Assess the financial impact of the firm's diverse service offerings.
- Identify the efficiency of cost management strategies within various components of EBITDA.
- Determine the effectiveness of current financial strategies in achieving long-term profitability and growth.
- Analyze the potential for market expansion and service optimization based on clear financial indicators.
- Evaluate the conversion of financial data into strategic business actions and decision-making.

### **Task**

As the data analyst for the consulting firm, my task was to dissect and translate the financial models coming from different country teams into a coherent, interactive Excel dashboard. The challenge was to uncover hidden patterns, provide a granular view of financial performance, and empower the finance team with actionable insights to drive revenue growth and cost efficiency.

### **Action**

I embarked on this analytical journey by synthesizing financial data sets into a comprehensive dashboard. Through meticulous data processing and the application of advanced Excel features like VLOOKUPS, INDEX, MATCH functions, Pivot Tables, Pivot Charts and slicers, I crafted interactive visualizations that revealed:

- The revenue and EBITDA trends over a ten-year span, offering a long-term perspective on the company's financial trajectory.
- A service profitability analysis that exposed the relative success of our offerings such as Accounting, Artificial Intelligence, and Marketing.
- A nuanced breakdown of EBITDA, uncovering insights into revenue, cost of goods sold (COGS), gross profit, overheads, and ultimately EBITDA itself, enabling precise cost control and profitability assessment.

![CleanShot 2024-01-29 at 19.14.58@2x.png](https://prod-files-secure.s3.us-west-2.amazonaws.com/dcb66927-c8f2-4a47-ac5c-f5ff4cd2d4fb/3c74b4ce-b95a-47b8-8daa-4821ed1b9a7c/CleanShot_2024-01-29_at_19.14.582x.png)

### **Results**

The deployed dashboard shed lights into the areas of improvement for the business, giving the finance team real actionable insight. 

The finance team's forecasting dashboard paints a concerning picture for the year 2024, flagging it as a year potentially marked by financial downturns for the consulting firm. A comparative analysis between the years 2023 and 2024 unveils that there is an anticipated sharp decline in revenue within the Marketing service line, plummeting from nearly £260,000 in 2023 to £200,000 in 2024. This downturn is not just a top-line revenue issue; the granular data points to an alarming swell in the cost of goods sold (COGS) and overhead expenses for 2024.

These red flags in the Marketing service area could signify various underlying issues such as reduced market demand, increased competition, or ineffective marketing strategies. The heightened COGS and overheads suggest that despite the falling revenue, the service line’s expenses are not being adjusted accordingly, which could lead to a strain on profit margins.

Given this stark forecast, the finance team is now in a position to sound the alarm and steer the company's focus toward overhauling the Marketing service's financial strategy. Corrective measures could potentially recover up to £100,000 in revenue for 2024, realigning the service line's performance with the firm's financial expectations.

 Next Steps

- Add a country filter and compare Service Profitability and EBITDA bridge by country
- Identify the years, services or countries (or a combination of these) where margins are optimised so that we can advise the leaders to replicate those optimal conditions